Move modals to separate components, add shared stylesheet
