+++ src/App.vue
@@ -111,88 +111,11 @@
     </div>
 
     <!-- Contact Modal -->
-    <b-modal id="contact-modal" @hide="handleContactModalHide" scrollable size="lg" centered>
-      <template v-slot:modal-title>
-        Contact {{ selectedRepresentative.name }}
-      </template>
-      <template v-slot:default>
-        <div id="email-tip">
-          <span class="icon"><v-icon name="regular/lightbulb" scale="1.2"/></span> Tip: Customize this template email to increase visibility from your representative
-        </div>
-        <div id="email-body">
-          To whom it may concern, <br> <br>
-          I am a concerned representative in your district writing today to voice my support and encouragement for the continued improvement of policing in our jurisdiction. I was shocked to see the brutal murder of George Floyd, and want to prevent such tragedies from occurring in the future. Over 1000 people are killed by police violence each year in America, and we can effectively reduce this violence by implementing research-based policy solutions.
-          <br> <br> 
-          These policies include: <br> 
-          - Increasing (or introducing) de-escalation training for police officers <br>
-          Currently, police officers only spend 8 hours on conflict de-escalation training, and this has been shown to reduce use of force. After introducing de-escalation training, the Dallas Police Department saw an 18% drop in use of force. <br> <br> 
-          - Increasing the use of Body Worn Cameras (BWCs) <br>
-          Introducing BWC’s also helps reduce force and helps increase accountability. Nearly every case where a police officer was charged with a crime for killing a civilian in 2015 relied on video evidence showing the officer's actions. <br> <br> 
-          - Reduce the size of the police force  <br>
-          Reducing the size of the police department and reinvesting that budget can help increase budgets for additional support systems for communities such as social workers, emergency medical technicians, conflict resolution specialists, restorative justice teams, and other community-based professionals. <br> <br> 
-          - Limiting the use of force used by police officers <br> 
-          Adding policies around the restricting the use of force by police officers has been shown to significantly reduce killings in those departments. Introducing these policies has been shown to reduce killings by up to 72%. <br> <br> 
-          As a taxpayer, I whole-heartedly support strengthening or introducing these policies in our jurisdiction. It’s clear that in the near future these sorts of training, as well as the use of BWCs, will be the standard for all jurisdictions in our country. I hope our jurisdiction can lead in determining best practices and policies for the safety of all in our community.
-          <br> <br> 
-          Thank you for your time and consideration! <br> <br> <br> 
-          Sources: <br>
-          - <a href="http://fivethirtyeight.com/features/a-chicago-cop-is-the-latest-to-be-charged-using-video-evidence/">http://fivethirtyeight.com/features/a-chicago-cop-is-the-latest-to-be-charged-using-video-evidence/</a> <br>
-          - <a href="https://www.joincampaignzero.org ">Campaign Zero </a> <br>
-          - <a href="http://govred.com/blog/deescalation-training-state-requirements/">http://govred.com/blog/deescalation-training-state-requirements/</a>
-        </div>
-      </template>
-      <template v-slot:modal-footer>
-        <div class="button-list">
-          <b-button
-            v-bind:href="'mailto:' + selectedRepresentative.emails[0] + '?subject=' + emailSubject + '&body=' + processedEmailBody"
-            v-if="selectedRepresentative.emails"
-            type="submit"
-            @click="fireEmailGAEvent"
-            class="modal-button"
-          >
-            <b-icon icon="envelope"></b-icon> Email
-          </b-button>
-          <b-button
-            v-bind:href="'tel:+1' + selectedRepresentative.phones"
-            v-if="selectedRepresentative.phones"
-            @click="firePhoneGAEvent"
-            class="modal-button"
-          >
-          <b-icon icon="phone"></b-icon>
-            Call
-          </b-button>
-
-          <b-popover ref="copy-paste-popover" target="copy-email-text-button" triggers="click blur" placement="top">
-            Email content copied!
-          </b-popover>
-
-          <b-button id="copy-email-text-button" class="modal-button" @click="copyToClipboard">
-            <content-copy-icon></content-copy-icon>
-          </b-button>
-
-          <b-button class="modal-button" :href="repTweetURL" v-if="repTweetURL" @click="tweetClicked" target="_blank">
-            <v-icon name="brands/twitter" scale="1.2"/>
-          </b-button>
-        </div>
-      </template>
-    </b-modal>
+    <contact-modal :selectedRepresentative="selectedRepresentative"></contact-modal>
 
     <!-- Thank You Modal -->
-    <b-modal id="thank-you-modal" content-class="modal-content" size="lg" hide-footer>
-      <template v-slot:modal-header="{ }">
-        <div>Thank you for joining this movement.</div>
-        <close-icon></close-icon>
-      </template>
+    <thank-you-modal></thank-you-modal>
 
-      <div id="thank-you-body">
-        <p>Together, with sustained pressure, we can make real change in policing.</p>
-        <p>Please spread the word using the links below. It will maximize the impact of this messaging campaign.</p>
-        <b-row class="social-icons-row" align-h="center">
-          <v-icon name="brands/facebook-square" scale="2.5" style="color: rgb(43, 43, 43)"/>
-          <v-icon name="brands/twitter" scale="2.5" style="color: rgb(43, 43, 43)"/>
-        </b-row>
-      </div>
-    </b-modal>
   </div>
 </template>
 
@@ -201,10 +124,15 @@
 import { getRepresentatives } from "./endpoints";
 import { emailSubject, emailBody } from './assets/email_template';
 import VueScrollTo from "vue-scrollto";
-import { extractSocialIdFromRep, makeTwitterLink, isAndroid, isSamsung } from './utils';
+import ContactModal from './components/ContactModal.vue';
+import ThankYouModal from './components/ThankYouModal.vue';
 
 export default {
   name: "App",
+  components: {
+    'contact-modal': ContactModal ,
+    'thank-you-modal': ThankYouModal
+  },
   created() {
     this.handleGoogleAnalyticsStartup();
     if(window.location.pathname === '/staging/') {
@@ -214,21 +142,6 @@ export default {
   mounted() {
     // focus input on page load, helpful for usability and screenreaders
     this.$refs.zipcode.$el.focus();
-  },
-  computed: {
-    repTweetURL: function() {
-      const id = extractSocialIdFromRep(this.selectedRepresentative, 'twitter');
-      if(!id) return null;
-      const text = `@${id} Please consider implementing policies outlined by: joincampaignzero.org (via: messageofprotest.com)`;
-      return makeTwitterLink(text, 'BlackLivesMatter');
-    },
-    processedEmailBody: function() {
-      let encoded = encodeURIComponent(emailBody);
-      // on all android devices except Samsung, default email client is Gmail, which uses '<br>' in
-      // mailto links for newlines instead of the standard '%0A'.  
-      if(isAndroid && !isSamsung) return encoded.replace(/%0A/g, '<br>');
-      return encoded;
-    },
   },
   methods: {
     autosubmitZip: async function() {
@@ -260,19 +173,6 @@ export default {
       this.$ga.event("select", "click", "selected a representative from the list");
       this.selectedRepresentative = rep;
     },
-    copyToClipboard: function() {
-
-      // once the "text has been copied" notification is shown, hide it after a few seconds
-      setTimeout(() => this.$refs['copy-paste-popover'].$emit('close'), 5000);
-
-      this.$ga.event("copy", "click", "user copied message text");
-      var range = document.createRange();
-      range.selectNode(document.getElementById("email-body"));
-      window.getSelection().removeAllRanges(); // clear current selection
-      window.getSelection().addRange(range); // to select text
-      document.execCommand("copy");
-      window.getSelection().removeAllRanges(); // to deselect
-    },
     /// TODO: most of this logic should be moved into the utils.js
     handleGoogleAnalyticsStartup: function() {
       const enable = this.$route.query.enable_google_analytics;
@@ -294,20 +194,8 @@ export default {
         this.$bvToast.toast(`To re-enable, go to ${basePath}=true`, { title: `Google Analytics Disabled` });
       }
     },
-    handleContactModalHide: function() {
-      this.$ga.event("contact modal hide", "click", "user clicked out of contact modal");
-    },
-    fireEmailGAEvent: function() {
-      this.$ga.event("email", "click", "user emailed representative");
-    }, 
-    firePhoneGAEvent: function() {
-      this.$ga.event("phone", "click", "user phoned representative");
-    },
     C0ButtonClicked: function() {
       this.$ga.event("C0", "click", "user clicked Campaign Zero link");
-    },
-    tweetClicked: function() {
-      this.$ga.event("tweet", "click", "user tweeted representative");
     },
   },
   data: () => ({
@@ -315,8 +203,6 @@ export default {
     zipcodeHasError: false,
     representatives: {},
     selectedRepresentative: { name: "Placeholder", emails: [], phones: [] },
-    emailSubject: encodeURIComponent(emailSubject),
-    copyBody: emailSubject,
     loading: false
   })
 };
@@ -324,50 +210,13 @@ export default {
 
 
 <style lang="scss">
-@import "bootstrap";
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:wght@700&family=Source+Sans+Pro:wght@300&display=swap");
-
-* {
-  font-family: "Source Sans Pro", sans-serif;
-}
-
-button,
-div,
-a {
-  border-radius: 0 !important;
-}
-
-a {
-  text-decoration: underline;
-  color: inherit;
-  transition-duration: 0.4s;
-}
-
-a:hover { 
-  color: #fce21b; 
-}
-
-.btn {
-  text-decoration: none;
-  transition-duration: 0.4s;
-}
-
-.btn:hover {
-  background-color: #fce21b;
-  border-color: #fce21b;
-  
-}
+@import 'styles.scss';
 
 #app {
   font-family: Avenir, Helvetica, Arial, sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   text-align: center;
-}
-
-body {
-  background-color: #121212;
-  color: white;
 }
 
 .button-main {
@@ -381,32 +230,9 @@ body {
   font-weight: 600;
 }
 
-#contact-modal {
-  color: black;
-}
-
 .footer-text {
   text-align: left !important;
   margin-bottom: 5px;
-}
-
-h1 {
-  font-size: 3rem;
-  font-family: "Montserrat", sans-serif;
-}
-
-h2 {
-  font-size: 2.5rem;
-  font-family: "Montserrat", sans-serif;
-}
-
-@include media-breakpoint-down(xs) {
-  h1 {
-    font-size: 2.5rem;
-  }
-  h2 {
-    font-size: 2rem;
-  }
 }
 
 .header {
@@ -448,12 +274,6 @@ h2 {
   }
 }
 
-.icon {
-  position: relative;
-  bottom: 2px;
-  padding-right: 4px;
-}
-
 @include media-breakpoint-down(xs) {
   .modal-dialog,
   .modal-content {
@@ -468,89 +288,6 @@ h2 {
   }
 }
 
-.rep-name {
-  font-weight: 600;
-}
-
-hr {
-  height: 2px;
-  background-color: white;
-  margin-top: 20px;
-}
-
-.modal-button {
-  float: left;
-  margin-right: 10px;
-
-  &:hover {
-    color: black;
-  }
-}
-
-// make buttons smaller on mobile to fit on one line
-@media (max-width: 350px) {
-  .modal-button {
-    padding: 0.25rem 0.5rem;
-    font-size: 0.9rem;
-  }
-}
-
-.modal-content {
-  color: black;
-  border: none;
-
-  .modal-body {
-    font-size: 20px;
-    margin: 0;
-    padding: 0;
-  }
-
-  #email-tip {
-    font-size: 16px !important;
-    background-color: #6c757d;
-    color: white;
-    padding: 10px 16px;
-  }
-
-  #email-body {
-    font-size: 20px;
-    padding: 16px;
-  }
-
-  .modal-header {
-    justify-content: center;
-    font-size: 24px;
-    padding-right: 20px;
-    text-align: center;
-
-    div {
-      font-family: "Montserrat", sans-serif;
-    }
-
-    .close-icon {
-      position: absolute;
-      right: 20px;
-    }
-  }
-
-  .modal-title {
-    font-family: "Montserrat", sans-serif;
-  }
-
-  .modal-footer {
-    justify-content: left;
-  }
-
-  .button-list {
-    display: flex;
-  }
-}
-
-.social-icons-row > * {
-  margin-left: 1rem;
-  margin-right: 1rem;
-}
-
 .rep-container {
   padding-bottom: 20px;
   padding-top: 25px;
@@ -561,6 +298,10 @@ hr {
     border-radius: 3px !important;
     box-shadow: 0px 0px 0px 1px #ffffff;
   }
+}
+
+.rep-name {
+  font-weight: 600;
 }
 
 .rep-section div:last-child > div.row > div.rep-container {
@@ -624,35 +365,5 @@ hr {
   right: 15px;
   position: absolute;
   border-width: 0.2rem;
-}
-
-.toast-body, .popover-body {
-  color: white !important;
-  background-color: #6c757d;
-}
-
-// hide up/down arrows on numbered inputs
-input[type="number"]::-webkit-inner-spin-button,
-input[type="number"]::-webkit-outer-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
-
-input[type="number"] {
-  -moz-appearance: textfield;
-}
-
-/* Change Autocomplete styles in Chrome*/
-input:-webkit-autofill,
-input:-webkit-autofill:hover,
-input:-webkit-autofill:focus,
-textarea:-webkit-autofill,
-textarea:-webkit-autofill:hover,
-textarea:-webkit-autofill:focus,
-select:-webkit-autofill,
-select:-webkit-autofill:hover,
-select:-webkit-autofill:focus {
-  -webkit-text-fill-color: white;
-  -webkit-box-shadow: 0 0 0px 1000px #000 inset;
 }
 </style>
