--- src/components/HelloWorld.vue
@@ -1,18 +1,0 @@
-<template>
-  <div>
-      HelloWorld object
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'HelloWorld',
-  props: {
-    msg: String
-  }
-}
-</script>
-
-<!-- Add "scoped" attribute to limit CSS to this component only -->
-<style scoped>
-</style>
